+++ jour01/job06/AverageOfFiveInts.java
@@ -12,8 +12,8 @@ public class AverageOfFiveInts {
         // Open the scanner
         Scanner myIntScanner = new Scanner (System.in);
 
-        // 3 times
-        for(int i = 0; i < 5; i++) {
+        // 5 times
+        for(byte i = 0; i < 5; i++) {
             
             // Say you expect print n° i using formatted string
             System.out.println(String.format("Entrez l'entier n° %s à comparer : ", i + 1));
