+++ jour01/job07/SumOfOneHundred.java
@@ -6,7 +6,7 @@ public class SumOfOneHundred {
 
         int sum = 0; 
 
-        for(int i = 1; i <= 100; i++) {
+        for(byte i = 1; i <= 100; i++) {
 
             sum += i;
         }
